refactor: each solution should be in its own package
Prevents an issue where multiple puzzles have a type that shouldn't be
shared between them. For example grid or point so each solution owns the
code written. If something should be shared an 'internal' directory will
be created and the code will be refactored. Also each test now lives
next to the package is it testings